# remove njit from simple functions