--- a/nbs/src/theta.ipynb
+++ b/nbs/src/theta.ipynb
@@ -962,7 +962,7 @@
     "            res[f'hi-{lv}'] = np.quantile(samples, max_q, axis=1)\n",
     "            \n",
     "    if obj.get('decompose', False):\n",
-    "        seas_forecast = _repeat_val_seas(obj['seas_forecast']['mean'], h=h, season_length=obj['m'])\n",
+    "        seas_forecast = _repeat_val_seas(obj['seas_forecast']['mean'], h=h)\n",
     "        for key in res:\n",
     "            if obj['decomposition_type'] == 'multiplicative':\n",
     "                res[key] = res[key] * seas_forecast\n",
@@ -1087,6 +1087,7 @@
   {
    "cell_type": "code",
    "execution_count": null,
+   "id": "cd79a3e6",
    "metadata": {},
    "outputs": [],
    "source": [