--- a/statsforecast/models.py
+++ b/statsforecast/models.py
@@ -1744,36 +1744,23 @@ def _ses_mse(alpha: float, x: np.ndarray) -> float:
     return mse
 
 
-@njit(nogil=NOGIL, cache=CACHE)
 def _ses_forecast(x: np.ndarray, alpha: float) -> Tuple[float, np.ndarray]:
     """One step ahead forecast with simple exponential smoothing."""
     forecast, _, fitted = _ses_fcst_mse(x, alpha)
     return forecast, fitted
 
 
-@njit(nogil=NOGIL, cache=CACHE)
 def _demand(x: np.ndarray) -> np.ndarray:
     """Extract the positive elements of a vector."""
     return x[x > 0]
 
 
-@njit(nogil=NOGIL, cache=CACHE)
 def _intervals(x: np.ndarray) -> np.ndarray:
     """Compute the intervals between non zero elements of a vector."""
-    y = []
-
-    ctr = 1
-    for val in x:
-        if val == 0:
-            ctr += 1
-        else:
-            y.append(ctr)
-            ctr = 1
-
-    return np.array(y)
+    nonzero_idxs = np.where(x != 0)[0]
+    return np.diff(nonzero_idxs + 1, prepend=0)
 
 
-@njit(nogil=NOGIL, cache=CACHE)
 def _probability(x: np.ndarray) -> np.ndarray:
     """Compute the element probabilities of being non zero."""
     return (x != 0).astype(np.int32)
@@ -1790,27 +1777,23 @@ def _optimized_ses_forecast(
     return forecast, fitted
 
 
-@njit(nogil=NOGIL, cache=CACHE)
 def _chunk_sums(array: np.ndarray, chunk_size: int) -> np.ndarray:
-    """Splits an array into chunks and returns the sum of each chunk."""
-    n = array.size
-    n_chunks = n // chunk_size
-    sums = np.empty(n_chunks)
-    for i, start in enumerate(range(0, n, chunk_size)):
-        sums[i] = array[start : start + chunk_size].sum()
-    return sums
+    """Splits an array into chunks and returns the sum of each chunk.
+
+    Incomplete chunks are discarded"""
+    n_chunks = array.size // chunk_size
+    n_elems = n_chunks * chunk_size
+    return array[:n_elems].reshape(n_chunks, chunk_size).sum(axis=1)
 
 # %% ../nbs/src/core/models.ipynb 118
-@njit(nogil=NOGIL, cache=CACHE)
 def _ses(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: bool,  # fitted values
     alpha: float,  # smoothing parameter
-):
+) -> Dict[str, np.ndarray]:
     fcst, _, fitted_vals = _ses_fcst_mse(y, alpha)
-    mean = _repeat_val(val=fcst, h=h)
-    fcst = {"mean": mean}
+    fcst = {"mean": _repeat_val(val=fcst, h=h)}
     if fitted:
         fcst["fitted"] = fitted_vals
     return fcst
@@ -2150,14 +2133,13 @@ class SimpleExponentialSmoothingOptimized(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 144
-@njit(nogil=NOGIL, cache=CACHE)
 def _seasonal_exponential_smoothing(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: bool,  # fitted values
     season_length: int,  # length of season
     alpha: float,  # smoothing parameter
-):
+) -> Dict[str, np.ndarray]:
     n = y.size
     if n < season_length:
         return {"mean": np.full(h, np.nan, np.float32)}
@@ -2168,7 +2150,7 @@ def _seasonal_exponential_smoothing(
         season_vals[i], fitted_vals[init_idx::season_length] = _ses_forecast(
             y[init_idx::season_length], alpha
         )
-    out = _repeat_val_seas(season_vals=season_vals, h=h, season_length=season_length)
+    out = _repeat_val_seas(season_vals=season_vals, h=h)
     fcst = {"mean": out}
     if fitted:
         fcst["fitted"] = fitted_vals
@@ -2278,9 +2260,7 @@ class SeasonalExponentialSmoothing(_TS):
         forecasts : dict
             Dictionary with entries `mean` for point predictions and `level_*` for probabilistic predictions.
         """
-        mean = _repeat_val_seas(
-            self.model_["mean"], season_length=self.season_length, h=h
-        )
+        mean = _repeat_val_seas(self.model_["mean"], h=h)
         res = {"mean": mean}
         if level is None:
             return res
@@ -2372,7 +2352,7 @@ def _seasonal_ses_optimized(
         season_vals[i], fitted_vals[init_idx::season_length] = _optimized_ses_forecast(
             y[init_idx::season_length], [(0.01, 0.99)]
         )
-    out = _repeat_val_seas(season_vals=season_vals, h=h, season_length=season_length)
+    out = _repeat_val_seas(season_vals=season_vals, h=h)
     fcst = {"mean": out}
     if fitted:
         fcst["fitted"] = fitted_vals
@@ -2478,9 +2458,7 @@ class SeasonalExponentialSmoothingOptimized(_TS):
         forecasts : dict
             Dictionary with entries `mean` for point predictions and `level_*` for probabilistic predictions.
         """
-        mean = _repeat_val_seas(
-            self.model_["mean"], season_length=self.season_length, h=h
-        )
+        mean = _repeat_val_seas(self.model_["mean"], h=h)
         res = {"mean": mean}
         if level is None:
             return res
@@ -2643,17 +2621,13 @@ class HoltWinters(AutoETS):
         return self.alias
 
 # %% ../nbs/src/core/models.ipynb 203
-@njit(nogil=NOGIL, cache=CACHE)
 def _historic_average(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: bool,  # fitted values
-):
-    mean = _repeat_val(val=y.mean(), h=h)
-    fcst = {"mean": mean}
+) -> Dict[str, np.ndarray]:
+    fcst = {"mean": _repeat_val(val=y.mean(), h=h)}
     if fitted:
-        # fitted_vals = np.full(y.size, np.nan, np.float32) # one-step ahead
-        # fitted_vals[1:] = y.cumsum()[:-1] / np.arange(1, y.size)
         fitted_vals = _repeat_val(val=y.mean(), h=len(y))
         fcst["fitted"] = fitted_vals
     return fcst
@@ -3046,16 +3020,15 @@ class Naive(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 233
-@njit(nogil=NOGIL, cache=CACHE)
 def _random_walk_with_drift(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: bool,  # fitted values
-):
+) -> Dict[str, np.ndarray]:
     slope = (y[-1] - y[0]) / (y.size - 1)
-    mean = slope * (1 + np.arange(h)) + y[-1]
+    mean = slope * (1 + np.arange(h, dtype=np.float32)) + y[-1]
     fcst = {
-        "mean": mean.astype(np.float32),
+        "mean": mean.astype(np.float32, copy=False),
         "slope": np.array([slope], dtype=np.float32),
         "last_y": np.array([y[-1]], dtype=np.float32),
     }
@@ -3333,9 +3306,7 @@ class SeasonalNaive(_TS):
         forecasts : dict
             Dictionary with entries `mean` for point predictions and `level_*` for probabilistic predictions.
         """
-        mean = _repeat_val_seas(
-            season_vals=self.model_["mean"], season_length=self.season_length, h=h
-        )
+        mean = _repeat_val_seas(season_vals=self.model_["mean"], h=h)
         res = {"mean": mean}
 
         if level is None:
@@ -3434,13 +3405,12 @@ class SeasonalNaive(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 264
-@njit(nogil=NOGIL, cache=CACHE)
 def _window_average(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: bool,  # fitted values
     window_size: int,  # window size
-):
+) -> Dict[str, np.ndarray]:
     if fitted:
         raise NotImplementedError("return fitted")
     if y.size < window_size:
@@ -3609,24 +3579,20 @@ class WindowAverage(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 276
-@njit(nogil=NOGIL, cache=CACHE)
 def _seasonal_window_average(
     y: np.ndarray,
     h: int,
     fitted: bool,
     season_length: int,
     window_size: int,
-):
+) -> Dict[str, np.ndarray]:
     if fitted:
         raise NotImplementedError("return fitted")
     min_samples = season_length * window_size
     if y.size < min_samples:
         return {"mean": np.full(h, np.nan, np.float32)}
-    season_avgs = np.zeros(season_length, np.float32)
-    for i, value in enumerate(y[-min_samples:]):
-        season = i % season_length
-        season_avgs[season] += value / window_size
-    out = _repeat_val_seas(season_vals=season_avgs, h=h, season_length=season_length)
+    season_avgs = y[-min_samples:].reshape(window_size, season_length).mean(axis=0)
+    out = _repeat_val_seas(season_vals=season_avgs, h=h)
     return {"mean": out}
 
 # %% ../nbs/src/core/models.ipynb 277
@@ -3723,9 +3689,7 @@ class SeasonalWindowAverage(_TS):
         forecasts : dict
             Dictionary with entries `mean` for point predictions and `level_*` for probabilistic predictions.
         """
-        mean = _repeat_val_seas(
-            season_vals=self.model_["mean"], season_length=self.season_length, h=h
-        )
+        mean = _repeat_val_seas(season_vals=self.model_["mean"], h=h)
         res = {"mean": mean}
         if level is None:
             return res
@@ -3989,7 +3953,6 @@ class ADIDA(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 302
-@njit(nogil=NOGIL, cache=CACHE)
 def _croston_classic(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
@@ -4344,12 +4307,11 @@ class CrostonOptimized(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 326
-@njit(nogil=NOGIL, cache=CACHE)
 def _croston_sba(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: bool,  # fitted values
-):
+) -> Dict[str, np.ndarray]:
     if fitted:
         raise NotImplementedError("return fitted")
     mean = _croston_classic(y, h, fitted)
@@ -4698,14 +4660,13 @@ class IMAPA(_TS):
         return res
 
 # %% ../nbs/src/core/models.ipynb 350
-@njit(nogil=NOGIL, cache=CACHE)
 def _tsb(
     y: np.ndarray,  # time series
     h: int,  # forecasting horizon
     fitted: int,  # fitted values
     alpha_d: float,
     alpha_p: float,
-):
+) -> Dict[str, np.ndarray]:
     if fitted:
         raise NotImplementedError("return fitted")
     if (y == 0).all():