Contributes to #753 by removing the njit decorator from functions that can be fast with just numpy. Also removes the `season_length` argument from `_repeat_val_seas`, since it can be inferred from the size of the `season_vals`.

Below are some benchmarks for the non-trivial functions, some increase its runtime but they're all in microseconds, so the impact wouldn't be noticeable (even with many series).

## _intervals
```python
x = np.random.randint(low=0, high=100, size=10_000).astype(np.float32)
%timeit _intervals(x)
```
| main | this PR |
| - | - |
| 43 µs | 38 µs |

## _chunk_sums
```python
x = np.random.rand(10_000)
%timeit _chunk_sums(x, 5)
%timeit _chunk_sums(x, 10)
%timeit _chunk_sums(x, 20)
```
| chunk_size | main | this PR |
| - | - | - |
| 5 | 10 µs | 35 µs |
| 10 | 7 µs | 23 µs |
| 20 | 6 µs | 13 µs |

## _seasonal_window_average
```python
x = np.random.rand(10_000)
h = 100
fitted = False
season_length = 52
%timeit _seasonal_window_average(x, h, fitted, season_length, 10)
%timeit _seasonal_window_average(x, h, fitted, season_length, 20)
%timeit _seasonal_window_average(x, h, fitted, season_length, 50)
```

| window_size | main | this PR |
| - | - | - |
| 10 | 6 µs   | 20 µs |
| 20 | 8 µs   | 20 µs |
| 50 | 15 µs | 19 µs |

## _repeat_val_seas
```python
seas_vals = np.random.rand(24)
%timeit _repeat_val_seas(seas_vals, 500)
```

| main | this PR |
| - | - |
| 3 µs | 5 µs |